fix: comprimir imagem

+++ src/helpers/file-helper.js
@@ -4,14 +4,16 @@ const sharp = require('sharp');
 
 exports.compressImage = async (file, size) => {
     const newPath = file.path.split('.')[0] + '.png';
-  
+    console.log(newPath)
 
     return sharp(file.path)
-        .resize({ 
+        .resize({
             width: 645,
-            height: 1024,})
+            height: 1024,
+        })
         .toFormat('png')
         .toBuffer()
+
         .then(data => {
 
             // Deletando o arquivo antigo
@@ -20,22 +22,32 @@ exports.compressImage = async (file, size) => {
 
                 // Um erro significa que a o arquivo não existe, então não tentamos apagar
                 if (!err) {
-                    
-                    //Se não houve erros, tentamos apagar
-                    fs.unlink(file.path, err => {
-                        // Não quero que erros aqui parem todo o sistema, então só vou imprimir o erro, sem throw.
-                        if(err) console.log(err)
-                    })
+                    if (file.filename.split('.')[1] == "png") {
+                        
+                    }else{
+
+                        fs.unlink(file.path, err => {
+                            // Não quero que erros aqui parem todo o sistema, então só vou imprimir o erro, sem throw.
+                            if (err) console.log(err)
+                        })
+                    }
                 }
+                //Se não houve erros, tentamos apagar
             });
-            
+
+
+
+
+
             //Agora vamos armazenar esse buffer no novo caminho
             fs.writeFile(newPath, data, err => {
-                if(err){
-                    
+                if (err) {
+
                     // Já aqui um erro significa que o upload falhou, então é importante que o usuário saiba.
-                    throw err;
+                    throw err
                 }
+
+
             });
 
             // Se o código chegou até aqui, deu tudo certo, então vamos retornar o novo caminho
